move handlers into a directory
